moves property-related stuff from DataServiceObserver to PropertyObserver
